fits: fix int overflow in fits-var heap size calculation
Fix a bug in the VariableFitsTableSerializer that calculated the
value of the THEAP header wrong.  It was using ints instead of
longs in one place.  This meant that variable-length
array FITS files longer than about 2Gb were written incorrectly.

+++ fits/src/main/uk/ac/starlink/fits/VariableFitsTableSerializer.java
@@ -74,7 +74,8 @@ public class VariableFitsTableSerializer extends StandardFitsTableSerializer {
     public Header getHeader() throws HeaderCardException {
         Header hdr = super.getHeader();
         long heapSize = getHeapSize();
-        long datEnd = hdr.getIntValue( "NAXIS1" ) * hdr.getIntValue( "NAXIS2" );
+        long datEnd =
+            hdr.getLongValue( "NAXIS1" ) * hdr.getLongValue( "NAXIS2" );
         long theap = ( ( datEnd + 2880 - 1 ) / 2880 ) * 2880;
         long gap = theap - datEnd;
         long pcount = gap + heapSize;
